Move collision rates and Hall parameter to collisions.py
